java: Avoid starting multiple JVMs in parallel using the same JAR
Java will randomly complain about the JAR being missing. This seems to
be some kind of JVM bug.

Workaround this by using a unique copy of the Tinkerforge.jar for each
JVM invocation.

+++ java/test_java_bindings.py
@@ -27,6 +27,7 @@ import sys
 import os
 import subprocess
 import shutil
+import random
 
 sys.path.append(os.path.split(os.getcwd())[0])
 import common
@@ -36,6 +37,23 @@ class JavaExamplesTester(common.Tester):
         common.Tester.__init__(self, 'java', '.java', root_dir, extra_paths=extra_paths)
 
     def test(self, cookie, path, extra):
+        # create unique copy of the Tinkerforge.jar to avoid Java from randomly
+        # complaining about the JAR being missing if Java is started multiple
+        # times concurrently
+        jar_path = '/tmp/tester/java/Tinkerforge.jar'
+        tries = 100
+
+        while os.path.exists(jar_path) and tries > 0:
+            r = int(round(random.random() * 10000000000))
+            jar_path = '/tmp/tester/java/Tinkerforge_{:010X}.jar'.format(r)
+            tries -= 1
+
+        if os.path.exists(jar_path):
+            self.handle_result(cookie, 1, 'cloud not create unique copy of Tinkerforge.jar')
+            return
+
+        shutil.copy('/tmp/tester/java/Tinkerforge.jar', jar_path)
+
         if extra:
             shutil.copy(path, '/tmp/tester/java')
             path = os.path.join('/tmp/tester/java', os.path.split(path)[1])
@@ -44,7 +62,7 @@ class JavaExamplesTester(common.Tester):
                 '-Xlint:all',
                 '-Werror',
                 '-cp',
-                '/tmp/tester/java/Tinkerforge.jar:.',
+                jar_path + ':.',
                 path]
 
         self.execute(cookie, args)
